test: cross-compile and run the suite under wine
make check-windows builds every suite with mingw-w64 and runs it under wine.
It stays out of test-all, because most machines have no cross compiler, but it
is the only thing that has ever compiled the _WIN32 branches, let alone run
them.

Two tests were asserting Linux behaviour rather than the library's:

  path_join inserts PATH_SEP, a backslash on Windows, so the expected strings
  now use the platform separator instead of a hardcoded slash.

  The logger's output was compared byte for byte from a file opened in text
  mode, where Windows turns every \n into \r\n. The helper opens in binary
  mode now, which is what a byte-exact comparison requires.

+++ tests/test_memory.c
@@ -41,7 +41,7 @@ TEST(arena_honours_over_alignment) {
         void *p = arena_alloc_aligned(&a, 32, align);
         if (!CHECK(p != NULL)) break;
         if (!CHECK((uintptr_t)p % align == 0))
-            printf("      align=%zu gave %p\n", align, p);
+            printf("      align=%lu gave %p\n", (unsigned long)align, p);
     }
     arena_free(&a);
 }
